refactoring packagenamen wahlvorschlaege -> wahlvorschlag
